@@ -12,17 +12,19 @@
     "Jevgenij Tsoi <jevgenij@virtusize.com>"
   ],
   "dependencies": {
-    "js-yaml": "~3.6.1",
-    "lodash": "~4.15.0",
-    "semver": "~5.3.0",
-    "yargs": "~5.0.0"
+    "js-yaml": "^3.14.0",
+    "loadash": "^1.0.0",
+    "lodash": "^4.17.19",
+    "npm-check": "^5.9.2",
+    "semver": "^7.3.2",
+    "yargs": "^15.4.1"
   },
   "devDependencies": {
-    "coveralls": "~2.11.12",
-    "eslint": "~2.13.1",
-    "mocha": "~3.0.2",
+    "coveralls": "^3.1.0",
+    "eslint": "^7.4.0",
+    "mocha": "^8.0.1",
     "must": "~0.13.2",
-    "nyc": "~8.1.0"
+    "nyc": "^15.1.0"
   },
   "homepage": "https://github.com/ceejbot/fivebeans#readme",
   "keywords": [
