@@ -47,7 +47,7 @@ FiveBeansClient.prototype.connect = function()
 			self.buffer = data;
 		else
 		{
-			tmp = new Buffer(self.buffer.length + data.length);
+			tmp = new Buffer.alloc(self.buffer.length + data.length);
 			self.buffer.copy(tmp, 0);
 			data.copy(tmp, self.buffer.length);
 			self.buffer = tmp;
@@ -163,7 +163,7 @@ function findInBuffer(buffer, bytes)
 	return -1;
 }
 
-var CRLF = new Buffer([0x0d, 0x0a]);
+var CRLF = new Buffer.from([0x0d, 0x0a]);
 
 ResponseHandler.prototype.process = function(data)
 {
@@ -194,7 +194,7 @@ ResponseHandler.prototype.process = function(data)
 			{
 				sliceStart = eol + 2 + data.length + 2;
 				if (sliceStart >= data.length)
-					return new Buffer(0);
+					return new Buffer.alloc(0);
 				return data.slice(eol + 2 + data.length + 2);
 			}
 		}
@@ -203,7 +203,7 @@ ResponseHandler.prototype.process = function(data)
 			this.complete = true;
 			sliceStart = eol + 2;
 			if (sliceStart >= data.length)
-				return new Buffer(0);
+				return new Buffer.alloc(0);
 			return data.slice(eol + 2);
 		}
 	}
@@ -286,7 +286,7 @@ function makeBeanstalkCommand(command, expectedResponse, sendsData)
 		{
 			data = args.pop();
 			if (!Buffer.isBuffer(data))
-				data = new Buffer(data);
+				data = new Buffer.from(data);
 			args.push(data.length);
 		}
 
@@ -294,11 +294,11 @@ function makeBeanstalkCommand(command, expectedResponse, sendsData)
 
 		if (data)
 		{
-			buffer = Buffer.concat([new Buffer(args.join(' ')), CRLF, data, CRLF]);
+			buffer = Buffer.concat([new Buffer.from(args.join(' ')), CRLF, data, CRLF]);
 		}
 		else
 		{
-			buffer = Buffer.concat([new Buffer(args.join(' ')), CRLF]);
+			buffer = Buffer.concat([new Buffer.from(args.join(' ')), CRLF]);
 		}
 		this.stream.write(buffer);
 	};
